Add support for rapidjson as the JSON library
nlohmann is still the default, but if `CPPDAP_JSON_DIR` points to rapidjson, then this will be used instead

Most of this was upstreamed from:
https://fuchsia-review.googlesource.com/c/third_party/github.com/google/cppdap/+/456566

+++ src/socket.cpp
@@ -302,7 +302,7 @@ std::shared_ptr<ReaderWriter> Socket::connect(const char* address,
 
       timeval tv;
       tv.tv_sec = microseconds / 1000000;
-      tv.tv_usec = microseconds - (tv.tv_sec * 1000000);
+      tv.tv_usec = microseconds - static_cast<uint32_t>(tv.tv_sec * 1000000);
       res = select(static_cast<int>(socket + 1), nullptr, &fdset, nullptr, &tv);
       if (res > 0 && !errored(socket) && setBlocking(socket, true)) {
         out = shared;
